@@ -10,14 +10,20 @@
         "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
     },
     "devDependencies": {
+        "autoprefixer": "^10.0.2",
         "axios": "^0.18.1",
-        "cross-env": "^5.1",
-        "laravel-mix": "^4.0.7",
-        "lodash": "^4.17.19",
+        "cross-env": "^5.2.1",
+        "laravel-mix": "^4.1.4",
+        "laravel-mix-tailwind": "^0.1.1",
+        "lodash": "^4.17.20",
+        "postcss": "^8.1.10",
         "resolve-url-loader": "^2.3.1",
-        "sass": "^1.15.2",
-        "sass-loader": "^7.1.0",
-        "vue": "^2.5.17",
-        "vue-template-compiler": "^2.6.10"
+        "sass": "^1.29.0",
+        "sass-loader": "^7.3.1",
+        "tailwindcss": "^2.0.1",
+        "vue": "^2.6.12",
+        "vue-template-compiler": "^2.6.12",
+        "no-more-design-system": "0.0.10",
+        "vue-router": "^3.4.9"
     }
 }
