@@ -1,4 +1,7 @@
 const mix = require('laravel-mix');
+// const tailwindcss = require('tailwindcss');
+
+require('laravel-mix-tailwind');
 
 /*
  |--------------------------------------------------------------------------
@@ -12,4 +15,5 @@ const mix = require('laravel-mix');
  */
 
 mix.js('resources/js/app.js', 'public/js')
-   .sass('resources/sass/app.scss', 'public/css');
+	.sass('resources/sass/app.scss', 'public/css')
+	.tailwind();
